feat: implement BATCH_010 tasks - compilation fixes & production hardening
BATCH_010 Implementation (12 tasks completed):

TASK-001: Fix SQLx Offline Mode for All Crates
- Converted 24 sqlx::query! macros to runtime sqlx::query() calls
- Files: mfa/mod.rs, progress.rs, recalculation.rs, entity_resolution.rs, query_log.rs

TASK-002: Fix Type Mismatches Across Crates
- HLCTimestamp::from_components type conversion (sync/repository.rs)
- String type mismatch fix (sync/publisher.rs)
- CommandType::Cast → CastTo enum variant (sync/command_router.rs)
- Redis TTL u64 cast (ingestion/deduplication.rs)
- Qdrant ListValue conversion (ingestion/qdrant.rs)
- f64/f32 quality_score casts (ingestion/repository.rs)

TASK-003: Fix totp-rs API Breaking Change
- Replaced Secret::generate_secret() with Secret::default()

TASK-004: Fix Rate Limiter ServiceResponse Type
- Implemented EitherBody<B, BoxBody> pattern for middleware

TASK-005: Fix AuditLogger get_logs() Signature
- Verified already uses correct AuditLogFilter struct

TASK-006: Add Missing HybridSearchService Field
- Added activity_producer: Option<Arc<KafkaActivityProducer>>

TASK-007: Fix Redis Never Type Fallback Warnings
- Added explicit type annotations to 36 Redis operations
- Files: storage.rs, email/service.rs, session.rs, rate_limit_config.rs

TASK-008: Implement MCP Server list_devices Tool
- New ListDevicesTool with DeviceInfo struct
- Database query for user devices
- Registered in handlers

TASK-009: Add MCP Server STDIO Transport
- Created transport/stdio.rs (479 lines)
- Added --stdio flag support in main.rs
- 13 unit tests passing

TASK-010: Register Missing Discovery HTTP Routes
- Search endpoints: POST /search, GET /autocomplete
- Analytics endpoints: GET /analytics
- Admin ranking endpoints: GET/PUT /admin/search/ranking/*
- Quality endpoints: GET /quality/report

TASK-011: Create Rust CI/CD Workflow
- .github/workflows/rust-ci.yml
- Jobs: check, test, build, docker
- SQLX_OFFLINE=true for CI

TASK-012: Create Prometheus Alert Rules
- config/prometheus/alerts.yml (28 alerts, 6 groups)
- Availability, latency, resources, business, database, Kafka alerts
- Updated prometheus.yml with rule_files

+++ crates/ingestion/src/webhooks/deduplication.rs
@@ -90,7 +90,7 @@ impl WebhookDeduplicator {
             .map_err(|e| WebhookError::RedisError(format!("Connection failed: {}", e)))?;
 
         // Store hash with TTL
-        let _: () = conn.set_ex(&key, webhook.timestamp.to_rfc3339(), self.ttl_seconds as usize).await
+        let _: () = conn.set_ex(&key, webhook.timestamp.to_rfc3339(), self.ttl_seconds as u64).await
             .map_err(|e| WebhookError::DeduplicationError(format!("Redis set failed: {}", e)))?;
 
         Ok(hash)
